+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/validation/impl/DynamicTemplateValidator.java
@@ -8,6 +8,7 @@ package org.hibernate.search.backend.elasticsearch.validation.impl;
 
 import org.hibernate.search.backend.elasticsearch.lowlevel.index.mapping.impl.DynamicTemplate;
 import org.hibernate.search.backend.elasticsearch.lowlevel.index.mapping.impl.PropertyMapping;
+import org.hibernate.search.backend.elasticsearch.reporting.impl.ElasticsearchValidationMessages;
 
 import com.google.gson.JsonElement;
