+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/validation/impl/ValidationErrorCollector.java
@@ -10,7 +10,8 @@ import java.util.ArrayDeque;
 import java.util.Deque;
 import java.util.Iterator;
 
-import org.hibernate.search.backend.elasticsearch.logging.impl.ElasticsearchEventContexts;
+import org.hibernate.search.backend.elasticsearch.reporting.impl.ElasticsearchEventContexts;
+import org.hibernate.search.backend.elasticsearch.reporting.impl.ElasticsearchValidationMessages;
 import org.hibernate.search.engine.reporting.spi.ContextualFailureCollector;
 import org.hibernate.search.engine.reporting.spi.EventContexts;
 import org.hibernate.search.util.common.AssertionFailure;
