--- mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/logging/impl/PojoEventContextMessages.java
@@ -1,26 +1,0 @@
-/*
- * Hibernate Search, full-text search for your domain model
- *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later
- * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
- */
-package org.hibernate.search.mapper.pojo.logging.impl;
-
-import org.hibernate.search.util.common.logging.impl.MessageConstants;
-
-import org.jboss.logging.Messages;
-import org.jboss.logging.annotations.Message;
-import org.jboss.logging.annotations.MessageBundle;
-
-/**
- * Message bundle for event contexts in the POJO mapper.
- */
-@MessageBundle(projectCode = MessageConstants.PROJECT_CODE)
-public interface PojoEventContextMessages {
-
-	PojoEventContextMessages INSTANCE = Messages.getBundle( PojoEventContextMessages.class );
-
-	@Message(value = "Schema management")
-	String schemaManagement();
-
-}
